+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/api/booklet/Controller.kt
@@ -15,6 +15,7 @@ import fr.sacane.jmanager.infrastructure.api.toDTO
 import fr.sacane.jmanager.infrastructure.api.toHttpResponse
 import fr.sacane.jmanager.infrastructure.api.transaction.TransactionResult
 import kotlinx.serialization.Serializable
+import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import java.time.LocalDate
@@ -34,7 +35,7 @@ class BookletController (
 
     @PostMapping
     fun saveBooklet(
-        @RequestBody bookletRequest: BookletBookingRequest
+        @Valid @RequestBody bookletRequest: BookletBookingRequest
     ): ResponseEntity<BookletInfoDTO> {
         LOGGER.info("Booking a new Booklet ${bookletRequest.label} starting at ${bookletRequest.amount}${bookletRequest.currency} for user ${currentUser.id}")
         return feature.save(